fix: proper gas calculation
